support-rag-bot/requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.12.0
